VIH-5432 refactor participant status list into base class
