+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/individual-participant-status-list/individual-participant-status-list.component.ts
@@ -1,67 +1,41 @@
-import { Component, Input, OnDestroy, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { AdalService } from 'adal-angular4';
-import { Subscription } from 'rxjs';
 import { ConsultationService } from 'src/app/services/api/consultation.service';
 import { VideoWebService } from 'src/app/services/api/video-web.service';
-import {
-    ConferenceResponse,
-    ConsultationAnswer,
-    EndpointStatus,
-    ParticipantResponse,
-    ParticipantStatus,
-    Role,
-    VideoEndpointResponse
-} from 'src/app/services/clients/api-client';
+import { ConsultationAnswer, ParticipantResponse, ParticipantStatus, VideoEndpointResponse } from 'src/app/services/clients/api-client';
 import { EventsService } from 'src/app/services/events.service';
 import { Logger } from 'src/app/services/logging/logger-base';
-import { AdminConsultationMessage } from 'src/app/services/models/admin-consultation-message';
 import { ConsultationMessage } from 'src/app/services/models/consultation-message';
 import { ParticipantStatusMessage } from 'src/app/services/models/participant-status-message';
 import { Hearing } from 'src/app/shared/models/hearing';
 import { Participant } from 'src/app/shared/models/participant';
 import { CaseTypeGroup } from 'src/app/waiting-space/models/case-type-group';
+import { WRParticipantStatusListDirective } from '../waiting-room-shared/wr-participant-list-shared.component';
 
 @Component({
     selector: 'app-individual-participant-status-list',
     templateUrl: './individual-participant-status-list.component.html',
     styleUrls: ['./individual-participant-status-list.component.scss']
 })
-export class IndividualParticipantStatusListComponent implements OnInit, OnDestroy {
-    @Input() conference: ConferenceResponse;
-
-    nonJugdeParticipants: ParticipantResponse[];
-    judge: ParticipantResponse;
-    panelMembers: ParticipantResponse[];
-    observers: ParticipantResponse[];
-    endpoints: VideoEndpointResponse[];
-
-    consultationRequestee: Participant;
-    consultationRequester: Participant;
-
-    adminConsultationMessage: AdminConsultationMessage;
-    eventHubSubscriptions$ = new Subscription();
-
+export class IndividualParticipantStatusListComponent extends WRParticipantStatusListDirective implements OnInit, OnDestroy {
     constructor(
-        private adalService: AdalService,
-        private consultationService: ConsultationService,
-        private eventService: EventsService,
-        private logger: Logger,
-        private videoWebService: VideoWebService
-    ) {}
+        protected adalService: AdalService,
+        protected consultationService: ConsultationService,
+        protected eventService: EventsService,
+        protected logger: Logger,
+        protected videoWebService: VideoWebService
+    ) {
+        super(adalService, consultationService, eventService, videoWebService, logger);
+    }
 
     ngOnInit() {
         this.consultationService.resetWaitingForResponse();
-        this.filterNonJudgeParticipants();
-        this.filterJudge();
-        this.filterPanelMembers();
-        this.filterObservers();
-        this.setupSubscribers();
+        this.initParticipants();
         this.endpoints = this.conference.endpoints;
     }
 
     ngOnDestroy(): void {
-        this.consultationService.clearOutoingCallTimeout();
-        this.eventHubSubscriptions$.unsubscribe();
+        this.executeTeardown();
     }
 
     setupSubscribers() {
@@ -101,27 +75,12 @@ export class IndividualParticipantStatusListComponent implements OnInit, OnDestr
         this.eventService.start();
     }
 
-    handleNoConsulationRoom() {
-        this.consultationService.displayNoConsultationRoomAvailableModal();
-    }
-
     handleParticipantStatusChange(message: ParticipantStatusMessage): void {
         const isCurrentUser = this.adalService.userInfo.userName.toLocaleLowerCase() === message.username.toLowerCase();
         if (isCurrentUser && message.status === ParticipantStatus.InConsultation) {
             this.closeAllPCModals();
         }
         this.filterNonJudgeParticipants();
-    }
-
-    async handleAdminConsultationMessage(message: AdminConsultationMessage) {
-        const requestee = this.conference.participants.find(x => x.username === message.requestedFor);
-        if (!this.isParticipantAvailable(requestee)) {
-            this.logger.info(`Ignoring request for private consultation from Video Hearings Team since participant is not available`);
-            return;
-        }
-        this.logger.info(`Incoming request for private consultation from Video Hearings Team`);
-        this.consultationRequestee = new Participant(requestee);
-        this.consultationService.displayAdminConsultationRequest();
     }
 
     canCallParticipant(participant: ParticipantResponse): boolean {
@@ -155,10 +114,6 @@ export class IndividualParticipantStatusListComponent implements OnInit, OnDestr
         }
 
         return this.isEndpointAvailable(endpoint);
-    }
-
-    getConsultationRequester(): ParticipantResponse {
-        return this.conference.participants.find(x => x.username.toLowerCase() === this.adalService.userInfo.userName.toLocaleLowerCase());
     }
 
     async beginCallWith(participant: ParticipantResponse): Promise<void> {
@@ -241,10 +196,6 @@ export class IndividualParticipantStatusListComponent implements OnInit, OnDestr
         }
     }
 
-    closeAllPCModals(): void {
-        this.consultationService.clearModals();
-    }
-
     private initConsultationParticipants(message: ConsultationMessage): void {
         const requester = this.conference.participants.find(x => x.username === message.requestedBy);
         const requestee = this.conference.participants.find(x => x.username === message.requestedFor);
@@ -256,37 +207,7 @@ export class IndividualParticipantStatusListComponent implements OnInit, OnDestr
         return participant.status === ParticipantStatus.Available ? 'Available' : 'Unavailable';
     }
 
-    isParticipantAvailable(participant: ParticipantResponse): boolean {
-        return participant.status === ParticipantStatus.Available;
-    }
-
-    isEndpointAvailable(endpoint: VideoEndpointResponse): boolean {
-        return endpoint.status === EndpointStatus.Connected;
-    }
-
-    private filterNonJudgeParticipants(): void {
-        this.nonJugdeParticipants = this.conference.participants.filter(
-            x => x.role !== Role.Judge && x.case_type_group !== CaseTypeGroup.OBSERVER && x.case_type_group !== CaseTypeGroup.PANEL_MEMBER
-        );
-    }
-
-    private filterJudge(): void {
-        this.judge = this.conference.participants.find(x => x.role === Role.Judge);
-    }
-
-    private filterPanelMembers(): void {
-        this.panelMembers = this.conference.participants.filter(x => x.case_type_group === CaseTypeGroup.PANEL_MEMBER);
-    }
-
-    private filterObservers(): void {
-        this.observers = this.conference.participants.filter(x => x.case_type_group === CaseTypeGroup.OBSERVER);
-    }
-
-    get getNumberParticipants() {
-        return this.nonJugdeParticipants.length + this.observers.length + this.panelMembers.length;
-    }
-
-    isCaseTypeNone(participant: ParticipantResponse): boolean {
-        return participant.case_type_group === 'None';
+    get getNumberPats() {
+        return this.nonJudgeParticipants.length + this.observers.length + this.panelMembers.length;
     }
 }
